refactoring & updates
